pass all files to ts

+++ src/hooks/useFile.ts
@@ -2,13 +2,14 @@ import { useState, useCallback } from 'react'
 
 import { toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
+import { FileUP } from '../entities/File'
 
 import { getFiles, updateService, saveFile, deleteFile } from '../services/files'
 
 export const useFile = () => {
   const [files] = useState([])
 
-  const saveInService = useCallback((file) => {
+  const saveInService = useCallback((file: FileUP) => {
     saveFile(file)
       .then(() => {
         toast.success('File saved successfully')
@@ -28,7 +29,11 @@ export const useFile = () => {
       })
   }, [])
 
-  const deleteFromService = ({ id }) => {
+  const deleteFromService = (id: string | number) => {
+    if (!id) {
+      toast.error('Error deleting file')
+      return
+    }
     deleteFile(id)
       .then(() => {
         toast.success('File deleted successfully')
